disable redirection on `try` mode
This commit changes how `hapi-auth-cookie` handles `try` mode.
before, if `hapi-auth-cookie` could not authenticate a user, and
the mode was `try` (i.e. auth is optinal), then the plugin would
redirect the user to the login form. This behaviour could be
overridden on a per-route basis using `route.config.plugins`.
however, now, if any request fails auth and the
mode is `try` then the user is not redirected. If redirection
is desired, then you can ask for it using `route.config.plugins`.